+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release Notes
 
+# Unreleased 
+
+### Added
+
+* Added Face Tracking in Selfie Capture to ensure only same face is used during capture
+
+### Changed
+
+* Removed `AntiFraud` response in `JobStatus` calls
+
+## 11.0.5 - July 2, 2025
+
+### Added
+
+* Added option to disable document auto capture in DocV and Enhanced DocV
+
 ## 11.0.4 - June 16, 2025
 
 ### Changed